Add response_type configuration option to Azure AD
The default value for response_type of `id_token code` requires that the app in the Azure AD portal  be configured to allow the implicit grant flow for id_token. If the `id_token` is not requested in the `:response_type`configuration the token is still returned and the authentication works without requiring the implicit grant configuration in the portal. 

This change allows the developer to override the response type. The default is currently to still have the `id_token code` value.