update to latest castore
I've created a fresh phoenix project (1.7.2) and added assent. Which resulted in:

```
❯ mix deps.get
Resolving Hex dependencies...
Resolution completed in 0.067s
Because "the lock" specifies castore 1.0.1 and every version of assent depends on castore ~> 0.1.0, the
And because your app depends on the lock, no version of assent is allowed.
So, because your app depends on assent ~> 0.2.2, version solving failed.
```